+++ database/migrations/2022_11_01_005508_create_detail_retur_pembelian_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
     public function up()
     {
         Schema::create('t_det_retur_pembelian', function (Blueprint $table) {
-            $table->integer('id')->autoIncrement();
-            $table->integer('id_retur_pembelian');
+            $table->bigInteger('id')->autoIncrement();
+            $table->bigInteger('id_retur_pembelian');
             $table->integer('id_barang');
             $table->integer('qty');
             $table->double('harga_beli');
